create new bid using jms and create singleton auction , user and bids repo

+++ ejb/src/main/java/lk/jiat/auction/ejb/remote/auctions/AuctionServices.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface AuctionServices {
     void addAuction(Auction auction);
     void removeAuction(Auction auction);
-    Auction getAuction(int auctionId);
+    Auction getAuction(Long auctionId);
     List<Auction> getAuctions();
 }
